+++ docs/source/change_log.rst
@@ -11,7 +11,7 @@ Version 1.x.x
 Version 1.12.0
 ^^^^^^^^^^^^^^
 
-*In Progress*
+Released :release:`1.12.0` on November 7th, 2018
 
 * Added support for users to set their email address for campaign alerts via email
 * Added additional plugin metadata fields for reference URLs and category classifiers
@@ -21,6 +21,10 @@ Version 1.12.0
    * There is now an option to update plugins in the menu
    * Plugins can ship with dedicated documentation in markdown files that will be displayed
    * The GUI no longer locks up while tasks like downloading plugins are taking place
+
+* Added the new ``fetch`` Jinja function and ``fromjson`` Jinja filter
+* Added ``campaign-alert-expired`` and ``campaign-expired`` server signals
+* Switched to using `Pipenv`_ to manage the environment and dependencies
 
 Version 1.11.0
 ^^^^^^^^^^^^^^
@@ -331,3 +335,5 @@ Released :release:`0.1.3` on June 4th, 2014
 
 * Jinja2 templates for both the client and server
 * API version checking to warn when the client and server versions are incompatible
+
+.. _Pipenv: https://pipenv.readthedocs.io/en/latest/